requirements.txt
# Medical Imaging DICOM Processing - Course Submission
# Requirements (minimal - no ML/API dependencies)

# Core Scientific Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# DICOM Processing
pydicom>=2.4.0

# Image Processing
pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning (only for train/test split utility)
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development (optional)
# jupyter>=1.0.0
# ipykernel>=6.0.0
